+++ benchmarks/bench_datasets.py
@@ -17,7 +17,7 @@ from functools import partial
 
 import numpy as np
 from astropy.modeling.models import Gaussian2D
-from bench_utils import print_environment, time_best
+from bench_helpers import print_environment, time_best
 
 from photutils.datasets import (apply_poisson_noise, make_4gaussians_image,
                                 make_100gaussians_image, make_gwcs,